config: update platforms to reflect swap_dev and cpu.h change

+++ Kernel/platform-coco2cart/config.h
@@ -16,7 +16,7 @@
 #define CONFIG_BANKS	1
 /* And swapping */
 
-extern unsigned int cocoswap_dev;
+extern uint16_t cocoswap_dev;
 
 #define SWAPDEV     cocoswap_dev    /* Uses part of IDE slice 0 or SD 0*/
 #define SWAP_SIZE   0x40	/* 32K in 512 byte blocks */
